+++ DownloadRouterConfig.py
@@ -18,99 +18,232 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 
-import datetime	# Required for date format
-import Exscript	# Required for SSH, queue & logging functionality
-import os		# Required to determine OS of host
+import argparse		# Required to read arguments from the command line
+import base64		# Required to decode password
+import datetime		# Required for date format
+import ConfigParser # Required for configuration file
+import Exscript		# Required for SSH, queue & logging functionality
+import re			# Required for REGEX operations
+import sys			# Required for printing without newline
+import os			# Required to determine OS of host
 
+from argparse					import ArgumentParser, RawDescriptionHelpFormatter
+from base64						import b64decode
+from ConfigParser				import ConfigParser
 from datetime                   import datetime
-from Exscript                   import Queue, Host, Logger
+from Exscript                   import Account, Queue, Host, Logger
 from Exscript.protocols 		import SSH2
 from Exscript.util.file			import get_hosts_from_file
 from Exscript.util.log          import log_to
 from Exscript.util.decorator    import autologin
 from Exscript.util.interact     import read_login
 from Exscript.util.report		import status,summarize
-from os							import mkdir, name, path, remove, system
+from re							import search, sub
+from sys						import stdout
+from os							import makedirs, name, path, system
+
+
+class Application:
+# This class was created to provide me with an easy way to update application
+# details across all my applications.  Also used to display information when
+# application is executed with "--help" argument.
+	author = "Aaron Melton <aaron@aaronmelton.com>"
+	date = "(2013-08-29)"
+	description = "Downloads the running-config of a Cisco router."
+	name = "DownloadRouterConfig.py"
+	version = "v2.2.1"
+	url = "https://github.com/aaronmelton/DownloadRouterConfig"
+
 
 logger = Logger()	# Log stuff
 @log_to(logger)		# Logging decorator; Must precede downloadRouterConfig!
 @autologin()		# Exscript login decorator; Must precede downloadRouterConfig!
-
 def downloadRouterConfig(job, host, socket):
-	socket.execute("terminal length 0")	# Disable user-prompt to page through config
-										# Exscript doesn't always recognize Cisco IOS
-										# for socket.autoinit() to work correctly
+# This function logs into each of the hosts in the routerFile and writes the 
+# contents of running-config to an outputFile.
+	# Define output filename based on hostname and date
+	outputFilename = backupDirectory+host.get_name()+'_Config_'+date+'.txt'	
 
-	socket.execute("show run")	# Show running config
+	# Check to see if outputFilename currently exists.  If it does, append an
+	# integer onto the end of the filename until outputFilename no longer exists
+	incrementOutputFilename = 1
+	while fileExist(outputFilename):
+		outputFilename = backupDirectory+host.get_name()+'_Config_'+date+'_'+str(incrementOutputFilename)+'.txt'
+		incrementOutputFilename = incrementOutputFilename + 1
 
-	configDirectory = ('configs_'+date+'/')	# Define directory to hold config files
-	if not path.exists(configDirectory): mkdir(configDirectory) # Create config file directory if it doesn't exist
+	with open(outputFilename, 'w') as outputFile:
+		try:
+			socket.execute("terminal length 0")		# Disable user-prompt to page through config
+													# Exscript doesn't always recognize Cisco IOS
+													# for socket.autoinit() to work correctly
+			socket.execute("show running-config")	# Show running config
+
+			# Write contents of running config to output file
+			outputFile.write(socket.response)
 		
-	outputFileName = host.get_name()+'_Config_'+date+'.txt'	# Define output filename based on hostname and date
-	outputFile = file(configDirectory+outputFileName, 'w')	# Open output file (will overwrite contents)
+			socket.send('exit\r')				# Send the "exit" command to log out of router gracefully
+			socket.close()						# Close SSH connection
 
-	outputFile.write(socket.response)	# Write contents of running config to output file
-	outputFile.close()					# Close output file
-	socket.send('exit\r')				# Send the "exit" command to log out of router gracefully
-	socket.close()						# Close SSH connection
+		# Exception: output file was not able to be opened
+		except IOError:
+			print
+			print "--> An error occurred opening "+outputFilename+"."
 
 def fileExist(fileName):
-# Check current path for existing file
+# This function checks the parent directory for the presence of a file
+# Returns true if found, false if not
+
 	try:
+		# If file can be opened, it must exist
 		with open(fileName, 'r') as openedFile:
-			# If file exists (can be opened), return true
-			return True
+			return True	# File found
+
+	# Exception: file cannot be opened, must not exist
 	except IOError:
-		# If file does not exists (can't be opened), return false
-		return 0
+		return False	# File NOT found
+
+
+# Check to determine if any arguments may have been presented at the command
+# line and generate help message for "--help" switch
+parser = ArgumentParser(
+    formatter_class=RawDescriptionHelpFormatter,description=(
+		Application.name+" "+Application.version+" "+Application.date+"\n"+
+		"--\n"+
+		"Description: "+Application.description+"\n\n"+
+		"Author: "+Application.author+"\n"+
+		"URL:    "+Application.url
+	))
+# Add additional argument to handle any optional configFile passed to application
+parser.add_argument("-c", "--config", dest="configFile", help="config file", default="settings.cfg", required=False)
+args = parser.parse_args()		# Set 'args' = input from command line
+configFile = args.configFile	# Set configFile = config file from command line OR 'settings.cfg'
 
 # Determine OS in use and clear screen of previous output
-system('cls' if name=='nt' else 'clear')
+if name == 'nt':	system("cls")
+else:				system("clear")
 
-print "Download Router Configuration v2.1.18"
-print "-------------------------------------"
-print
+# PRINT PROGRAM BANNER
+print Application.name+" "+Application.version+" "+Application.date
+print "-"*(len(Application.name+Application.version+Application.date)+2)
 
-# Define file with router IP Addresses or Hostnames
-routerFile = 'routers.txt'
 
-# Check for existence of routerFile; If exists, continue with program
-if fileExist(routerFile):
-	# Define 'date' variable for use in the output filename
-	date = datetime.now()	# Determine today's date
-	date = date.strftime('%Y%m%d')	# Format date as YYYYMMDD
-
-	# Read hosts from specified file & remove duplicate entries, set protocol to SSH2
-	hosts = get_hosts_from_file(routerFile,default_protocol='ssh2',remove_duplicates=True)
-	userCreds = read_login()	# Prompt the user for his name and password
-
-	print # Required for pretty spacing. :)
-
-	queue = Queue(verbose=1, max_threads=4)	# Minimal message from queue, 4 threads
-	queue.add_account(userCreds)			# Use supplied user credentials
-	queue.run(hosts, downloadRouterConfig)	# Create queue using provided hosts
-	queue.shutdown()						# End all running threads and close queue
-
-	print status(logger)	# Print current % status of operation to screen
-
-	logFile = open('status_'+date+'.log', 'w')	# Open 'status.log' file
-	logFile.write(summarize(logger))			# Write results of program to file
-	logFile.close()								# Close 'status.log' file
-
-# If routerFile does not exist, create example and exit
-else:
-	# Attempt to open routerFile to create an example
+# START PROGRAM
+try:
+# Try to open configFile
+	with open(configFile, 'r'):	pass
+	
+except IOError:
+# Except if configFile does not exist, create an example configFile to work from
 	try:
-		with open (routerFile, 'w') as exampleFile:
-			# Write example IP Addresses or Hostnames to routerFile
-			exampleFile.write("192.168.1.1\n192.168.1.2\nRouterA\nRouterB\nRouterC\netc...")
-			# Print error message
-			print "Required file "+routerFile+" not found; One has been created for you."
+		with open (configFile, 'w') as exampleFile:
+			print
+			print "--> Config file not found; Creating "+configFile+"."
+			exampleFile.write("## DownloadRouterConfig.py CONFIGURATION FILE ##\n#\n")
+			exampleFile.write("[account]\n#password is base64 encoded! Plain text passwords WILL NOT WORK!\n#Use website such as http://www.base64encode.org/ to encode your password\nusername=\npassword=\n#\n")
+			exampleFile.write("[DownloadRouterConfig]#variable=C:\path\\to\\filename.ext\nrouterFile=routers.txt\nlogFileDirectory=\nbackupDirectory=\n")
+			exampleFile.write("# See http://knipknap.github.io/exscript/api/Exscript.Queue-class.html#__init__\n# for information on verbose and thread settings\n")
+			exampleFile.write("# Recommend verboseOutput = 1\nverboseOutput=1\n# Recommend maxThreads > 2\nmaxThreads=5\n")
+
+	# Exception: configFile could not be created
+	except IOError:
+		print
+		print "--> An error occurred creating the example "+configFile+"."
+
+finally:
+# Finally, using the provided configFile (or example created), pull values
+# from the config and login to the router(s)
+	config = ConfigParser(allow_no_value=True)
+	config.read(configFile)
+	username = config.get('account', 'username')
+	password = config.get('account', 'password')
+	routerFile = config.get('DownloadRouterConfig', 'routerFile')
+	logFileDirectory = config.get('DownloadRouterConfig', 'logFileDirectory')
+	backupDirectory = config.get ('DownloadRouterConfig', 'backupDirectory')
+	verboseOutput = config.get ('DownloadRouterConfig', 'verboseOutput')
+	maxThreads = config.get ('DownloadRouterConfig', 'maxThreads')
+
+	# If logFileDirectory does not contain trailing backslash, append one
+	if logFileDirectory != '':
+		if logFileDirectory[-1:] != "\\":
+			logFileDirectory = logFileDirectory+"\\"
+			if not path.exists(logFileDirectory): makedirs(logFileDirectory)
+	# If backupDirectory does not contain trailing backslash, append one
+	if backupDirectory != '':
+		if backupDirectory[-1:] != "\\":
+			backupDirectory = backupDirectory+"\\"
+			if not path.exists(backupDirectory): makedirs(backupDirectory)
+
+	# Error checking for verboseOutput & maxThreads
+	if int(verboseOutput) not in range(0,5):	verboseOutput = 1
+	if int(maxThreads) not in range(1,100):		maxThreads = 2
+
+	if fileExist(routerFile):
+		# Define 'date' variable for use in the output filename
+		date = datetime.now()	# Determine today's date
+		date = date.strftime('%Y%m%d')	# Format date as YYYYMMDD
+	
+		if username == '':				# If username is blank
+			print
+			account = read_login()		# Prompt the user for login credentials
+
+		elif password == '':			# If password is blank
+			print
+			account = read_login()		# Prompt the user for login credentials
+
+		else:							# Else use username/password from configFile
+			account = Account(name=username, password=b64decode(password))
+	
+		# Read hosts from specified file & remove duplicate entries, set protocol to SSH2
+		hosts = get_hosts_from_file(routerFile, default_protocol='ssh2', remove_duplicates=True)
+		
+		print
+		
+		# Verbose & # threads taken from configFile
+		queue = Queue(verbose=int(verboseOutput), max_threads=int(maxThreads), stderr=(open(os.devnull, 'w')))
+		queue.add_account(account)				# # Use supplied user credentials
+		queue.run(hosts, downloadRouterConfig)	# Create queue using provided hosts
+		queue.shutdown()						# End all running threads and close queue
+	
+		print status(logger)	# Print current % status of operation to screen
+
+		# Define log filename based on date
+		logFilename = logFileDirectory+'BackupStatus_'+date+'.log'
+
+		# Check to see if logFilename currently exists.  If it does, append an
+		# integer onto the end of the filename until logFilename no longer exists
+		incrementLogFilename = 1
+		while fileExist(logFilename):
+			logFilename = logFileDirectory+'BackupStatus_'+date+'_'+str(incrementLogFilename)+'.log'
+			incrementLogFilename = incrementLogFilename + 1
+
+		# Write log results to logFile
+		with open(logFilename, 'w') as outputLogFile:
+			try:
+				outputLogFile.write(summarize(logger))
+
+			# Exception: router file was not able to be opened
+			except IOError:
+				print
+				print "--> An error occurred opening "+logFileDirectory+logFile+"."
+
+	else: # if fileExist(routerFile):
+	# Else if no routerFile exists, create a sample one and quit the program.
+		try:
+			with open (routerFile, 'w') as exampleFile:
+				print
+				print "--> Router file not found; Creating "+routerFile+"."
+				print "    Edit this file and restart the application."
+				exampleFile.write("## VRFSearchAndBackup.py ROUTER FILE ##\n#\n")
+				exampleFile.write("#Enter a list of hostnames or IP Addresses, one per line.\n#For example:\n")
+				exampleFile.write("192.168.1.1\n192.168.1.2\nRouterA\nRouterB\nRouterC\netc...")
+		
+		# Exception: file could not be created
+		except IOError:
+			print
+			print "--> Required file "+routerFile+" not found; An error has occurred creating "+routerFile+"."
 			print "This file must contain a list, one per line, of Hostnames or IP addresses the"
 			print "application will then connect to download the running-config."
-	# If unable to write file for whatever reason, just print error message
-	except IOError:
-		# Print error message
-		print "Required file "+routerFile+" not found."
-		print "This file must contain a list, one per line, of Hostnames or IP addresses the"
-		print "application will then connect to download the running-config."
+
+print
+print "--> Done."
+raw_input()	# Pause for user input.